Fix for the UI bug regarding checkboxes
If a user where to click the area where the checkbox were to appear even if it isn't visible it still registers the click. Therefore if you display the checkbox after clicking that specific area, it will show a blue checkbox. This has no harm to the adding and deleting logic of files but still was unintentional there the bug fix.